fc-ceph: factor out cryptsetup wrapper with opinionated tunables
